% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{calc_obj}
\alias{calc_obj}
\title{Objective function}
\usage{
calc_obj(z, y)
}
\arguments{
\item{z}{A vector of length L}

\item{y}{A vector of length L}
}
\value{
A scalar
}
\description{
Compute the objective function for optimization: the mean square error between the approximation z and the observed mean y over the L values of the control parameter.
}
\details{
This function expects that some mean state, y, can be or has been computed for a system. The function then tests the approximation, z, against the true mean state, y. The objective function is the sum of squared errors between z and y, normalized by the length L times the grand mean (the mean of y). The value that it returns is the approximation error for the given approximation z to the true mean state y.
}
\examples{
y <- 1:10
z <- sample(1:10, 10)
calc_obj(z, y)
}
