% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievers.R
\name{convenience}
\alias{convenience}
\alias{get_vs}
\alias{get_error}
\alias{get_ks}
\alias{get_ws}
\title{Retrieve information from lists of sentinel node sets}
\usage{
get_vs(dl)

get_error(dl)

get_ks(dl)

get_ws(dl)
}
\arguments{
\item{dl}{The list of node sets}
}
\value{
A matrix (vs, ks, ws if not NULL) or numeric vector (error).
}
\description{
Convenience functions to retrieve information from lists of sentinel node sets, typically the output of make_dataset().
}
