% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-datasets.R
\name{make_dataset}
\alias{make_dataset}
\title{Make a list of node sets}
\usage{
make_dataset(
  ntrials,
  ns.type = c("opt", "rand", "fixed", "constr", "quant", "comm", "knnconstr"),
  ncores = 1,
  ...
)
}
\arguments{
\item{ntrials}{The number of independent realizations of the node set type}

\item{ns.type}{One of 'opt', 'rand', 'fixed', 'constr', 'quant', or 'comm'}

\item{ncores}{Passed to mclapply()}

\item{...}{Arguments to be passed to the selector function.}
}
\value{
A list of node sets.
}
\description{
Make a list of node sets (and some additional information) according to a specified algorithm.
}
\details{
This function generates several realizations of a given algorithm, optionally with optimized node weights. See the manual page for each function for the required arguments.

Algorithms select nodes by the following methods: 'opt', combinatorial simulated annealing; 'rand', uniformly at random; 'fixed', uniformly at random, except that a provided vector of nodes specifies the degree sequence; 'constr', uniformly random, except that the largest 5\\% of nodes, by degree, are removed from consideration; 'quant', as constr, but additionally selecting one node at random from n 'bins' of the degree distribution; 'comm', as constr, but preferring to select nodes from different communities.
}
\seealso{
\code{\link[=select_optimized]{select_optimized()}}
}
