% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{quadoptm}
\alias{quadoptm}
\title{Quadratic programming function}
\usage{
quadoptm(vs, y, Y)
}
\arguments{
\item{vs}{The selected nodes (i.e., the columns of Y)}

\item{y}{The mean state at each level of the bifurcation parameter}

\item{Y}{The state of each node (in columns) at each level of the bifurcation parameter (rows)}
}
\value{
A vector of weights for the nodes in \code{vs}.
}
\description{
Optimization of node weights by quadratic programming, using ROI and the qpoases algorithm.
}
