% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{update_vs}
\alias{update_vs}
\title{Select a new node}
\usage{
update_vs(vs, g, y, Y, optimize_weights)
}
\arguments{
\item{vs}{The selected nodes (i.e., the columns of Y)}

\item{g}{The network.}

\item{y}{The mean state at each level of the bifurcation parameter}

\item{Y}{The state of each node (in columns) at each level of the bifurcation parameter (rows)}

\item{optimize_weights}{Logical for whether or not to optimize node weights.}
}
\value{
A numeric vector of node indices
}
\description{
Given a starting sample of nodes, select a random node from the node set and replace it with a random node not in the node set. Primarily for use inside optim().
}
