% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{obj_fn}
\alias{obj_fn}
\title{Calculate the approximated mean state for a sample of nodes}
\usage{
obj_fn(vs, y, Y, optimize_weights = FALSE, ws = NULL, ...)
}
\arguments{
\item{vs}{A sample of nodes}

\item{y}{The ground truth mean state}

\item{Y}{A matrix}

\item{optimize_weights}{Logical for whether or not to optimize node weights.}

\item{ws}{A vector of node weights, if already obtained}

\item{...}{Container for arguments passed from optim()}
}
\value{
A numeric vector of approximated system states at each value of the bifurcation parameter
}
\description{
Given a sample of nodes, compute the objective function: the weighted or unweighted mean state of the sample of nodes, normalized appropriately. Can be used on its own or inside optim().
}
\details{
The primary purpose of this function is to handle various features of the provided node set and return the approximation error for that node set. By default this function will

For the vector \code{vs}, it is safer to use the numeric indices of the graph to avoid confusion with vertex names.
}
\seealso{
\code{\link[=calc_obj]{calc_obj()}}
}
